+++ src/util.hpp
@@ -34,7 +34,7 @@ namespace Sass {
     return m;
   }
 
-  double round(double val, size_t precision = 0);
+  double round(double val, int precision = 0);
   double sass_strtod(const char* str);
   const char* safe_str(const char *, const char* = "");
   sass::string string_to_output(const sass::string& str);
